Custom deserializer + lab 4 completed

+++ OOP_Lab1/OOP_Lab1/Models/Gun.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -9,8 +10,8 @@ namespace OOP_Lab1.Models
     [Serializable]
     public class Gun : Module, ICopied<Gun>
     {
-        private double caliber;
-        public double Caliber
+        private int caliber;
+        public int Caliber
         {
             get { return caliber; }
             set
@@ -20,8 +21,8 @@ namespace OOP_Lab1.Models
             }
         }
 
-        private double rapidity;
-        public double Rapidity
+        private int rapidity;
+        public int Rapidity
         {
             get { return rapidity; }
             set
